Refactor: Move types to common
